requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
Pillow==10.4.0
torch>=2.1.0
torchvision>=0.16.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
matplotlib==3.9.0
streamlit==1.37.0
